refactor: rename detection module to vision and update package names, imports, and references accordingly
