+++ intake/pom.xml
@@ -26,7 +26,7 @@
         </dependency>
         <dependency>
             <groupId>pt.estga</groupId>
-            <artifactId>detection</artifactId>
+            <artifactId>vision</artifactId>
             <version>0.0.1-SNAPSHOT</version>
         </dependency>
         <dependency>
