+++ intake/src/main/java/pt/estga/intake/listeners/SubmissionEnrichmentListener.java
@@ -8,8 +8,8 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
-import pt.estga.detection.DetectionResult;
-import pt.estga.detection.DetectionService;
+import pt.estga.vision.DetectionResult;
+import pt.estga.vision.DetectionService;
 import pt.estga.file.application.MediaService;
 import pt.estga.intake.events.SubmissionSubmittedEvent;
 import pt.estga.intake.repositories.MarkEvidenceSubmissionRepository;
