+++ vision/pom.xml
@@ -10,7 +10,7 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>detection</artifactId>
+    <artifactId>vision</artifactId>
 
     <properties>
         <maven.compiler.source>25</maven.compiler.source>
