+++ vision/src/main/java/pt/estga/vision/config/DetectionAutoConfiguration.java
@@ -1,4 +1,4 @@
-package pt.estga.detection.config;
+package pt.estga.vision.config;
 
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.persistence.autoconfigure.EntityScan;
@@ -7,8 +7,8 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.web.client.RestTemplate;
 
 @AutoConfiguration
-@ComponentScan("pt.estga.detection")
-@EntityScan("pt.estga.detection")
+@ComponentScan("pt.estga.vision")
+@EntityScan("pt.estga.analysis")
 public class DetectionAutoConfiguration {
 
     @Bean
